// backend/src/generated/models/createMessageRequest.ts
/**
 * Pooper API
 * API for the Pooper chat application. Provides endpoints for chat, MCP, tools, analytics, realtime, config, prompt, and preview services. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateMessageRequest {
    'content': string;
    'userId': string;
    'username': string;
    'conversationId'?: string;
    'parentId'?: string;
    'type'?: CreateMessageRequest.TypeEnum = CreateMessageRequest.TypeEnum.text;
    'metadata'?: { [key: string]: any | undefined; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "content",
            "baseName": "content",
            "type": "string"
        },
        {
            "name": "userId",
            "baseName": "userId",
            "type": "string"
        },
        {
            "name": "username",
            "baseName": "username",
            "type": "string"
        },
        {
            "name": "conversationId",
            "baseName": "conversationId",
            "type": "string"
        },
        {
            "name": "parentId",
            "baseName": "parentId",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "CreateMessageRequest.TypeEnum"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any | undefined; }"
        }    ];

    static getAttributeTypeMap() {
        return CreateMessageRequest.attributeTypeMap;
    }
}

export namespace CreateMessageRequest {
    export enum TypeEnum {
        text = <any> 'text',
        code = <any> 'code',
        system = <any> 'system'
    }
}