// backend/src/generated/models/enhancePromptRequest.ts
/**
 * Pooper API
 * API for the Pooper chat application. Provides endpoints for chat, MCP, tools, analytics, realtime, config, prompt, and preview services. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { EnhancePromptRequestOptions } from './enhancePromptRequestOptions';

export class EnhancePromptRequest {
    'prompt': string;
    'options'?: EnhancePromptRequestOptions;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "prompt",
            "baseName": "prompt",
            "type": "string"
        },
        {
            "name": "options",
            "baseName": "options",
            "type": "EnhancePromptRequestOptions"
        }    ];

    static getAttributeTypeMap() {
        return EnhancePromptRequest.attributeTypeMap;
    }
}